requirements.txt
mcp>=1.3.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
starlette>=0.40.0
watchfiles>=0.24.0
